scraper/requirements.txt
httpx
pydantic>=2.0
tenacity
python-dotenv
